+++ src/automations/formula_functions/ScriptFunctions.ts
@@ -1,15 +1,16 @@
-import { Notice, TFile } from "obsidian";
+import { TFile } from "obsidian";
 import { IGenerateObject } from "automations/IGenerateObject";
 import { get_tfiles_from_folder } from "helpers/FileManagement";
 import { LocalSettings } from "cdm/SettingsModel";
 import { LOGGER } from "services/Logger";
+import { AutomationError, showDBError } from "errors/ErrorTypes";
 
-export class UserScriptFunctions implements IGenerateObject {
+export class ScriptFunctions implements IGenerateObject {
     constructor(private config: LocalSettings) { }
 
-    async generate_user_script_functions(
+    async generate_script_functions(
     ): Promise<Map<string, Function>> {
-        const user_script_functions: Map<string, Function> = new Map();
+        const script_functions: Map<string, Function> = new Map();
         const files =
             get_tfiles_from_folder(
                 this.config.formula_folder_path,
@@ -21,18 +22,25 @@ export class UserScriptFunctions implements IGenerateObject {
 
         for (const file of files) {
             if (file.extension.toLowerCase() === "js") {
-                await this.load_user_script_function(
-                    file,
-                    user_script_functions
-                );
+                try {
+                    await this.load_script_function(
+                        file,
+                        script_functions
+                    );
+                } catch (e) {
+                    showDBError({
+                        error: AutomationError.LoadFormulas.error,
+                        solution: `check your ${file.path} js file code`
+                    }, e);
+                }
             }
         }
-        return user_script_functions;
+        return script_functions;
     }
 
-    async load_user_script_function(
+    async load_script_function(
         file: TFile,
-        user_script_functions: Map<string, Function>
+        script_functions: Map<string, Function>
     ): Promise<void> {
         let req = (s: string) => {
             return window.require && window.require(s);
@@ -48,20 +56,20 @@ export class UserScriptFunctions implements IGenerateObject {
         const formula_function = exp['default'] || mod.exports;
 
         if (!formula_function) {
-            const msg = `Failed to load user script ${file.path}. No exports detected.`;
+            const msg = `Failed to load script ${file.path}. No exports detected.`;
             LOGGER.error(msg);
             return;
         }
         if (!(formula_function instanceof Function)) {
-            const msg = `Failed to load user script ${file.path}. Default export is not a function.`
+            const msg = `Failed to load script ${file.path}. Default export is not a function.`
             LOGGER.error(msg);
             return;
         }
-        user_script_functions.set(`${file.basename}`, formula_function);
+        script_functions.set(`${file.basename}`, formula_function);
     }
 
     async generate_object(): Promise<Record<string, unknown>> {
-        const user_script_functions = await this.generate_user_script_functions();
-        return Object.fromEntries(user_script_functions);
+        const script_functions = await this.generate_script_functions();
+        return Object.fromEntries(script_functions);
     }
 }