+++ src/components/styles/RowTemplateStyles.ts
@@ -11,9 +11,7 @@ const CustomTemplateSelectorStyles: StylesConfig<any, true, GroupBase<any>> = {
             ...styles,
             color: 'var(--text-normal)',
             backgroundColor: 'var(--background-primary)',
-            fontSize: '12px',
             textAlign: 'left',
-            width: 'auto',
         }
     },
     control: (styles) => {
@@ -21,9 +19,7 @@ const CustomTemplateSelectorStyles: StylesConfig<any, true, GroupBase<any>> = {
             ...styles,
             color: 'var(--text-normal)',
             backgroundColor: 'var(--background-primary)',
-            fontSize: '12px',
             textAlign: 'left',
-            minWidth: '17rem',
         }
     }
 }