+++ src/helpers/StylesHelper.ts
@@ -1,4 +1,4 @@
-import { DatabaseCore, MetadataLabels } from "helpers/Constants";
+import { COLUMN_ALIGNMENT_OPTIONS, DatabaseCore, MetadataLabels } from "helpers/Constants";
 
 /**
  * Wrap the classname of css elements
@@ -39,4 +39,8 @@ export function getLabelHeader(input: string) {
         }
     });
     return labelCandidate === undefined ? input : labelCandidate[1];
+}
+
+export function getAlignmentClassname(alignment: string) {
+    return alignment === undefined ? COLUMN_ALIGNMENT_OPTIONS.LEFT : alignment;
 }